[Refactor] 1차 리팩토링 - 컴포넌트 분리 및 커스텀 훅 분리, 파일정리
