+++ src/redux/index.ts
@@ -1,4 +1,3 @@
 export * from './api';
 export * from './config';
 export * from './modules';
-export * from './reduxType.d';
