+++ src/types/async.d.ts
@@ -1,3 +1,5 @@
+import * as Type from 'axios';
+
 // React-Redux를 위한 타입지정
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
@@ -10,4 +12,8 @@ export interface DecodeToken {
   gender: string;
   exp: number;
   iat: number;
+}
+
+export interface MyAxiosRequestConfig extends Type.AxiosRequestConfig {
+  headers: Type.AxiosRequestHeaders;
 }
