+++ src/types/index.ts
@@ -1,6 +1,6 @@
-export * from './mock'
 export * from './responseType'
 export * from './post'
 export * from './auth'
 export * from './styled'
 export * from './props'
+export * from './async'
