--- src/types/mock.d.ts
@@ -1,10 +1,0 @@
-export interface Login {
-  email: string;
-  password: string;
-}
-
-export interface Info {
-  success?: boolean;
-  info: string;
-  error?: null;
-}
